requirements.txt
streamlit==1.43.2
supabase==2.15.1
openai==1.68.2
python-dotenv==1.1.0
bcrypt==4.3.0
pandas==2.2.3
plotly==5.24.1
